pyproject.toml
[tool.poetry]
name = "cars"
version = "0.1.0"
description = ""
authors = ["wumpscut0"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
djangorestframework = "^3.15.2"

[tool.poetry.group.other]
black = "^24.8.0"

[tool.poetry.group.docs.dependencies]
drf-spectacular = "^0.27.2"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
